requirements.txt
gspread
google-auth
python-dotenv
pillow
requests
python-dateutil
pymupdf
pytesseract
opencv-python-headless
